use features and mixins to apply algo-specific logic
