+++ h2o-bindings/bin/custom/python/gen_aggregator.py
@@ -1,4 +1,5 @@
 rest_api_version = 99
+supervised_learning = False
 
 
 def class_extensions():
