+++ h2o-bindings/bin/custom/python/gen_defaults.py
@@ -8,3 +8,15 @@ def update_param(name, param):
         return param
     return None  # means no change
 
+
+# extra logic common to several algos
+overrides = dict(
+    checkpoint=dict(
+        setter="""
+assert_is_type(checkpoint, None, str, H2OEstimator)
+key = checkpoint.key if isinstance(checkpoint, H2OEstimator) else checkpoint
+self._parms["{sname}"] = key
+"""
+    ),
+)
+
