+++ h2o-bindings/bin/custom/python/gen_gbm.py
@@ -1,3 +1,14 @@
+options = dict(
+    verbose=True,
+    model_extensions=[
+        'h2o.model.extensions.ScoringHistory',
+        'h2o.model.extensions.VariableImportance',
+        'h2o.model.extensions.FeatureInteraction',
+        'h2o.model.extensions.Trees',
+    ],
+)
+
+
 def update_param(name, param):
     if name == 'distribution':
         param['values'].remove('ordinal')
