+++ h2o-bindings/bin/custom/python/gen_glrm.py
@@ -1,3 +1,6 @@
+supervised_learning = False
+
+
 doc = dict(
     __class__="""
 Builds a generalized low rank model of a H2O dataset.
