+++ h2o-bindings/bin/custom/python/gen_kmeans.py
@@ -1,3 +1,6 @@
+supervised_learning = False
+
+
 doc = dict(
     __class__="""Performs k-means clustering on an H2O dataset.""",
 )
