+++ h2o-bindings/bin/custom/python/gen_pca.py
@@ -1,3 +1,6 @@
+supervised_learning = False
+
+
 def class_extensions():
     def init_for_pipeline(self):
         """
