+++ h2o-bindings/bin/custom/python/gen_svd.py
@@ -1,4 +1,6 @@
 rest_api_version = 99
+supervised_learning = False
+
 
 def class_extensions():
     def init_for_pipeline(self):
