+++ h2o-py/h2o/estimators/estimator_base.py
@@ -16,6 +16,7 @@ from h2o.base import Keyed
 from h2o.exceptions import H2OValueError, H2OResponseError
 from h2o.frame import H2OFrame
 from h2o.job import H2OJob
+from h2o.utils.mixin import load_ext, mixin
 from h2o.utils.shared_utils import quoted
 from h2o.utils.typechecks import assert_is_type, is_type, numeric, FunctionType
 from ..model.autoencoder import H2OAutoEncoderModel
@@ -54,6 +55,8 @@ class H2OEstimator(ModelBase):
 
     Because H2OEstimator instances are instances of ModelBase, these objects can use the H2O model API.
     """
+    
+    supervised_learning = None  # overridden in implementation
 
     def start(self, x, y=None, training_frame=None, offset_column=None, fold_column=None,
               weights_column=None, validation_frame=None, **params):
@@ -80,7 +83,6 @@ class H2OEstimator(ModelBase):
                    validation_frame=validation_frame,
                    **params)
 
-
     def join(self):
         """Wait until job's completion."""
         self._future = False
@@ -89,7 +91,6 @@ class H2OEstimator(ModelBase):
         self._job = None
         model_json = h2o.api("GET /%d/Models/%s" % (self._rest_version, model_key))["models"][0]
         self._resolve_model(model_key, model_json)
-
 
     def train(self, x=None, y=None, training_frame=None, offset_column=None, fold_column=None,
               weights_column=None, validation_frame=None, max_runtime_secs=None, ignored_columns=None,
@@ -187,7 +188,6 @@ class H2OEstimator(ModelBase):
         job.poll()
         return H2OSegmentModels(job.dest_key)
 
-
     def _train(self, parms, verbose=False):
         assert_is_type(verbose, bool)
 
@@ -208,7 +208,6 @@ class H2OEstimator(ModelBase):
         job.poll(poll_updates=self._print_model_scoring_history if verbose else None)
         model_json = h2o.api("GET /%d/Models/%s" % (rest_ver, job.dest_key))["models"][0]
         self._resolve_model(job.dest_key, model_json)
-
 
     def _make_parms(self, x=None, y=None, training_frame=None, offset_column=None, fold_column=None,
                     weights_column=None, validation_frame=None, max_runtime_secs=None, ignored_columns=None,
@@ -233,25 +232,14 @@ class H2OEstimator(ModelBase):
             self._verify_training_frame_params(offset_column, fold_column, weights_column, validation_frame)
             training_frame = self.training_frame if has_default_training_frame else None
     
-        algo = self.algo
-        if verbose and algo not in ["drf", "gbm", "deeplearning", "xgboost"]:
-            raise H2OValueError("Verbose should only be set to True for drf, gbm, deeplearning, and xgboost models")
+        if verbose and not self._options_.get('verbose', False):
+            raise H2OValueError("Verbose mode is not available for %s" % self.__class__.__name__)
         parms = self._parms.copy()
-        if algo=="pca" and "k" not in parms.keys():
-            parms["k"] = 1
-        if "__class__" in parms:  # FIXME: hackt for PY3
-            del parms["__class__"]
-        is_auto_encoder = bool(parms.get("autoencoder"))
-        is_supervised = not(is_auto_encoder or algo in {"aggregator", "pca", "svd", "kmeans", "glrm", "word2vec", "isolationforest", "generic"})
-    
         names = training_frame.names if training_frame is not None else []
         ncols = training_frame.ncols if training_frame is not None else 0
         types = training_frame.types if training_frame is not None else {}
     
-        if "checkpoint" in parms and isinstance(parms["checkpoint"], H2OEstimator):
-            parms["checkpoint"] = parms["checkpoint"].key
-    
-        if is_supervised:
+        if self.supervised_learning:
             if y is None: y = "response"
             if is_type(y, int):
                 if not (-ncols <= y < ncols):
@@ -266,6 +254,7 @@ class H2OEstimator(ModelBase):
             # it. This way an unsupervised model can be used as a step in
             # sklearn's pipeline.
             y = None
+            self._estimator_type = "unsupervised"
     
         if override_default_training_frame:
             assert_is_type(y, str, None)
@@ -309,15 +298,6 @@ class H2OEstimator(ModelBase):
         if model_id is not None:
             parms["model_id"] = model_id
     
-        # Step 2
-        is_auto_encoder = "autoencoder" in parms and parms["autoencoder"]
-        is_unsupervised = is_auto_encoder or self.algo in {"aggregator", "pca", "svd", "kmeans", "glrm", "word2vec", "isolationforest"}
-        if is_auto_encoder and y is not None:
-            raise ValueError("y should not be specified for autoencoder.")
-        if not is_unsupervised and y is None and self.algo not in ["generic"]:
-            raise ValueError("Missing response")
-    
-        # Step 3
         if override_default_training_frame:
             parms["training_frame"] = training_frame
             offset = parms["offset_column"]
@@ -352,10 +332,8 @@ class H2OEstimator(ModelBase):
             raise H2OValueError("r2 cannot be used as an early stopping_metric yet.  Check this JIRA https://0xdata.atlassian.net/browse/PUBDEV-5381 for progress.")
         return parms
 
-
     def _get_rest_version(self, parms):
         return parms.pop("_rest_version") if "_rest_version" in parms else 3
-
 
     def _print_model_scoring_history(self, job, bar_progress=0):
         """
@@ -374,7 +352,6 @@ class H2OEstimator(ModelBase):
         except AttributeError:  # To catch NoneType error if scoring history is not available
             print("Scoring History is not available yet...")
 
-
     @staticmethod
     def _keyify(item):
         if isinstance(item, Keyed):
@@ -384,18 +361,13 @@ class H2OEstimator(ModelBase):
         else:
             return item
 
-
     def _resolve_model(self, model_id, model_json):
         metrics_class, model_class, metrics_class_valid = H2OEstimator._metrics_class(model_json)
         m = model_class()
         m._id = model_id
         m._model_json = model_json
-        if (model_json["algo"]=="glm") and self.HGLM:
-            m._have_pojo = False
-            m._have_mojo = False
-        else:
-            m._have_pojo = model_json.get('have_pojo', True)
-            m._have_mojo = model_json.get('have_mojo', True)
+        m._have_pojo = model_json.get('have_pojo', True)
+        m._have_mojo = model_json.get('have_mojo', True)
         m._metrics_class = metrics_class
         m._metrics_class_valid = metrics_class_valid
         m._parms = self._parms
@@ -403,6 +375,7 @@ class H2OEstimator(ModelBase):
         m._start_time = model_json.get('output', {}).get('start_time', None)
         m._end_time = model_json.get('output', {}).get('end_time', None)
         m._run_time = model_json.get('output', {}).get('run_time', None)
+        m._options_ = self._options_
 
         if model_id is not None and model_json is not None and metrics_class is not None:
             # build Metric objects out of each metrics
@@ -423,36 +396,10 @@ class H2OEstimator(ModelBase):
             # build a useful dict of the params
             for p in m._model_json["parameters"]:
                 m.parms[p["name"]] = p
-        H2OEstimator.mixin(self, model_class)
+                
+        extensions = [load_ext(ext) for ext in self._options_.get('model_extensions', [])]
+        mixin(self, model_class, *extensions)
         self.__dict__.update(m.__dict__.copy())
-
-
-    # TODO: replace with a property which is overriden in subclasses
-    def _compute_algo(self):
-        name = self.__class__.__name__
-        if name == "H2ODeepLearningEstimator": return "deeplearning"
-        if name == "H2OAutoEncoderEstimator": return "deeplearning"
-        if name == "H2OGradientBoostingEstimator": return "gbm"
-        if name == "H2OGeneralizedLinearEstimator": return "glm"
-        if name == "H2OGeneralizedLowRankEstimator": return "glrm"
-        if name == "H2OKMeansEstimator": return "kmeans"
-        if name == "H2ONaiveBayesEstimator": return "naivebayes"
-        if name == "H2ORandomForestEstimator": return "drf"
-        if name == "H2OXGBoostEstimator": return "xgboost"
-        if name == "H2OCoxProportionalHazardsEstimator": return "coxph"
-        if name == "H2OGeneralizedAdditiveEstimator": return "gam"
-        if name == "H2OIsolationForestEstimator": return "isolationforest"
-        if name in ["H2OPCA", "H2OPrincipalComponentAnalysisEstimator"]: return "pca"
-        if name in ["H2OSVD", "H2OSingularValueDecompositionEstimator"]: return "svd"
-        if name == "H2ORuleFitEstimator": return "rulefit"
-
-
-    @staticmethod
-    def mixin(obj, cls):
-        for name in cls.__dict__:
-            if name.startswith("__") and name.endswith("__"): continue
-            if not isinstance(cls.__dict__[name], types.FunctionType): continue
-            obj.__dict__[name] = cls.__dict__[name].__get__(obj)
 
 
     #------ Scikit-learn Interface Methods -------
@@ -484,7 +431,6 @@ class H2OEstimator(ModelBase):
         self.train(x, y, training_frame, **params)
         return self
 
-
     def get_params(self, deep=True):
         """
         Obtain parameters for this estimator.
@@ -503,7 +449,6 @@ class H2OEstimator(ModelBase):
                 out.update((key + "__" + k, val) for k, val in deep_items)
             out[key] = value
         return out
-
 
     def set_params(self, **parms):
         """
@@ -570,51 +515,6 @@ class H2OEstimator(ModelBase):
         if valid_metrics_class is None:
             valid_metrics_class = metrics_class
         return [metrics_class, model_class, valid_metrics_class]
-
-    def convert_H2OXGBoostParams_2_XGBoostParams(self):
-        """
-        In order to use convert_H2OXGBoostParams_2_XGBoostParams and convert_H2OFrame_2_DMatrix, you must import
-        the following toolboxes: xgboost, pandas, numpy and scipy.sparse.
-
-        Given an H2OXGBoost model, this method will generate the corresponding parameters that should be used by
-        native XGBoost in order to give exactly the same result, assuming that the same dataset
-        (derived from h2oFrame) is used to train the native XGBoost model.
-
-        Follow the steps below to compare H2OXGBoost and native XGBoost:
-
-         1. Train the H2OXGBoost model with H2OFrame trainFile and generate a prediction:
-
-          - h2oModelD = H2OXGBoostEstimator(\*\*h2oParamsD) # parameters specified as a dict()
-          - h2oModelD.train(x=myX, y=y, training_frame=trainFile) # train with H2OFrame trainFile
-          - h2oPredict = h2oPredictD = h2oModelD.predict(trainFile)
-
-         2. Derive the DMatrix from H2OFrame:
-         
-          - nativeDMatrix = trainFile.convert_H2OFrame_2_DMatrix(myX, y, h2oModelD)
-
-         3. Derive the parameters for native XGBoost:
-         
-          - nativeParams = h2oModelD.convert_H2OXGBoostParams_2_XGBoostParams()
-
-         4. Train your native XGBoost model and generate a prediction:
-         
-          - nativeModel = xgb.train(params=nativeParams[0], dtrain=nativeDMatrix, num_boost_round=nativeParams[1])
-          - nativePredict = nativeModel.predict(data=nativeDMatrix, ntree_limit=nativeParams[1]
-
-         5. Compare the predictions h2oPredict from H2OXGBoost, nativePredict from native XGBoost.
-
-        :return: nativeParams, num_boost_round
-        """
-        import xgboost as xgb
-
-        nativeParams = self._model_json["output"]["native_parameters"]
-        nativeXGBoostParams = dict()
-
-        for (a,keyname,keyvalue) in nativeParams.cell_values:
-            nativeXGBoostParams[keyname]=keyvalue
-        paramsSet = self.full_parameters
-
-        return nativeXGBoostParams, paramsSet['ntrees']['actual_value']
 
     def _check_and_save_parm(self, parms, parameter_name, parameter_value):
         """
