+++ h2o-py/h2o/estimators/gam.py
@@ -31,6 +31,8 @@ class H2OGeneralizedAdditiveEstimator(H2OEstimator):
     """
 
     algo = "gam"
+    supervised_learning = True
+    _options_ = {'model_extensions': ['h2o.model.extensions.ScoringHistoryGLM']}
 
     @deprecated_params({'Lambda': 'lambda_'})
     def __init__(self,
