+++ h2o-py/h2o/estimators/generic.py
@@ -19,6 +19,7 @@ class H2OGenericEstimator(H2OEstimator):
     """
 
     algo = "generic"
+    supervised_learning = False
 
     def __init__(self,
                  model_id=None,  # type: Optional[Union[None, str, H2OEstimator]]
