+++ h2o-py/h2o/estimators/rulefit.py
@@ -21,6 +21,7 @@ class H2ORuleFitEstimator(H2OEstimator):
     """
 
     algo = "rulefit"
+    supervised_learning = True
 
     def __init__(self,
                  model_id=None,  # type: Optional[Union[None, str, H2OEstimator]]
