+++ h2o-py/h2o/estimators/svd.py
@@ -19,6 +19,7 @@ class H2OSingularValueDecompositionEstimator(H2OEstimator):
     """
 
     algo = "svd"
+    supervised_learning = False
 
     def __init__(self,
                  model_id=None,  # type: Optional[Union[None, str, H2OEstimator]]
